+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/ClientManager.java
@@ -143,6 +143,7 @@ public class ClientManager {
         } catch (Exception e) {
             Voicechat.LOGGER.error("Failed to change voice chat port: {}", e.getMessage());
         }
+        Minecraft.getInstance().gui.getChat().addMessage(Component.translatable("message.voicechat.server_port", server.getPort()));
     }
 
     public void checkMicrophonePermissions() {